+++ KalikoCMS.Admin/Admin/Content/EditPage.aspx.cs
@@ -35,6 +35,7 @@ namespace KalikoCMS.Admin.Content {
             GetQueryStringValues();
 
             SaveButton.Click += SaveButtonEventHandler;
+            PublishButton.Click += PublishButtonEventHandler;
 
             _controls = new List<PropertyEditorBase>();
 
@@ -112,7 +113,7 @@ namespace KalikoCMS.Admin.Content {
         }
 
         private void LoadFormForExistingPage() {
-            var cmsPage = PageFactory.GetPage(_pageId);
+            var cmsPage = PageFactory.GetWorkingCopy(_pageId);
 
             _pageName = cmsPage.PageName;
             PageHeader.Text = _pageName;
@@ -143,6 +144,7 @@ namespace KalikoCMS.Admin.Content {
             }
         }
 
+
         private void SaveButtonEventHandler(object sender, EventArgs e) {
             if(IsDataValid) {
                 SaveData();
@@ -152,7 +154,25 @@ namespace KalikoCMS.Admin.Content {
                     Feedback.Visible = true;
                 }
                 else {
-                    ShowMessage(Feedback, String.Format("Page <b>{0}</b> has been saved!", _pageName));
+                    ShowMessage(Feedback, String.Format("A working copy of <b>{0}</b> has been saved! Remember to publish this version once it's complete.", _pageName));
+                }
+            }
+            else {
+                ShowError(Feedback, "One or more errors occured!");
+            }
+        }
+
+        private void PublishButtonEventHandler(object sender, EventArgs e) {
+            if(IsDataValid) {
+                var page = SaveData();
+                page.Publish();
+
+                if (_pageTypeId > 0) {
+                    Feedback.Text = string.Format("<script>parent.$('.notifications.top-right').notify({{ type: 'info', message: \"<i class=\\\"icon-flag\\\"></i> Page <b>{0}</b> has been created and published!!\", fadeOut: {{ enabled: true, delay: 5000 }}}}).show();parent.refreshTreeNode('{1}','{2}');document.location = '{3}?id={2}';</script>", _pageName, _parentId, _pageId, Request.Path);
+                    Feedback.Visible = true;
+                }
+                else {
+                    ShowMessage(Feedback, String.Format("Page <b>{0}</b> has been published!", _pageName));
                 }
             }
             else {
@@ -181,29 +201,33 @@ namespace KalikoCMS.Admin.Content {
             }
         }
 
-        private void SaveData() {
+        private EditablePage SaveData() {
             if(_pageId == Guid.Empty) {
-                SaveDataForNewPage();
+                return SaveDataForNewPage();
             }
             else {
-                SaveDataForExistingPage();
+                return SaveDataForExistingPage();
             }
         }
 
-        private void SaveDataForNewPage() {
+        private EditablePage SaveDataForNewPage() {
             var parent = PageFactory.GetPage(_parentId);
             var editablePage = parent.CreateChildPage(_pageTypeId);
             SavePropertiesForPage(editablePage);
             _pageId = editablePage.PageId;
             OldPageUrlSegment.Value = editablePage.UrlSegment;
+
+            return editablePage;
         }
 
-        private void SaveDataForExistingPage() {
-            var cmsPage = PageFactory.GetPage(_pageId);
+        private EditablePage SaveDataForExistingPage() {
+            var cmsPage = PageFactory.GetWorkingCopy(_pageId);
             _parentId = cmsPage.ParentId;
             var editablePage = cmsPage.MakeEditable();
             SavePropertiesForPage(editablePage);
             OldPageUrlSegment.Value = editablePage.UrlSegment;
+
+            return editablePage;
         }
 
         private void SavePropertiesForPage(EditablePage editablePage) {
@@ -216,9 +240,9 @@ namespace KalikoCMS.Admin.Content {
 
             PageUrlSegmentWasChanged = HandlePageUrlSegment(editablePage);
 
-            foreach (PropertyEditorBase propertyControl in _controls) {
-                string propertyName = propertyControl.PropertyName;
-                PropertyData propertyValue = propertyControl.PropertyValue;
+            foreach (var propertyControl in _controls) {
+                var propertyName = propertyControl.PropertyName;
+                var propertyValue = propertyControl.PropertyValue;
 
                 editablePage.SetProperty(propertyName, propertyValue);
             }
@@ -232,7 +256,7 @@ namespace KalikoCMS.Admin.Content {
         }
 
         public bool PageUrlSegmentWasChanged { get; set; }
-        
+
         public string CurrentPageId { get { return _pageId.ToString(); } }
 
         private bool HandlePageUrlSegment(EditablePage editablePage) {